Remove need to build tutorial plugin for types
