+++ plugins/tutorial/src/index.ts
@@ -4,21 +4,7 @@ import type {
   Plugin,
 } from '@docusaurus/types';
 
-export interface Options {
-  title: string;
-  description: string;
-  tags: string[];
-  source?: string;
-  preview?: string;
-}
-
-export interface DefaultOptions {
-  id: string;
-  route: string;
-}
-
-export type UserOptions = Options & Partial<DefaultOptions>;
-export type NormalizedOptions = Options & DefaultOptions;
+import type { NormalizedOptions, UserOptions } from './types';
 
 export default async function pluginTutorial(
   context: LoadContext,
